08-08-2016
Got the two lines in the correct position.

Was able to draw the arrow pointing down with CSS and it looks to be in the correct position.

The footer at the bottom of the page still doesn't look correct and I think it has something to do
with the overall height of the bottom page. Changing that does effect the position of the top nav
bar, so it might be one of those things where I have to tweek all three to just get the look that
I want.