08-16-16
Vincent suggested to start the CSS over and work with the smallest size first.

What I found is pretty intresting. There is a lot of positioning happening that are abosolute and relative and they are all working
on top of one another. It's all a matter of getting the parent child relations correct to get things to work better.

There is also a positioning issue with the menu as well right now and it is effecting how the top parallax is working.